Separate most scraper functions to scraper.js
